CPM-714: Move GetProductUuids to Enrichment\Product\API
