+++ src/Akeneo/Pim/Enrichment/Product/back/Infrastructure/Symfony/Resources/config/queries.yml
@@ -20,7 +20,7 @@ services:
         arguments:
             - '@database_connection'
 
-    Akeneo\Pim\Enrichment\Product\Domain\Query\GetProductUuids:
+    Akeneo\Pim\Enrichment\Product\API\Query\GetProductUuids:
         class: Akeneo\Pim\Enrichment\Product\Infrastructure\Query\Cache\LRUCacheGetProductUuids
         arguments:
             - '@Akeneo\Pim\Enrichment\Product\Infrastructure\Query\SqlGetProductUuids'
