+++ src/Akeneo/Pim/Enrichment/Product/back/Infrastructure/Symfony/Resources/config/validators.yml
@@ -10,7 +10,7 @@ services:
         arguments:
             - '@Akeneo\Pim\Enrichment\Product\Domain\Query\GetCategoryCodes'
             - '@Akeneo\Pim\Enrichment\Category\API\Query\GetOwnedCategories'
-            - '@Akeneo\Pim\Enrichment\Product\Domain\Query\GetProductUuids'
+            - '@Akeneo\Pim\Enrichment\Product\API\Query\GetProductUuids'
         tags:
             - { name: validator.constraint_validator }
 
@@ -19,7 +19,7 @@ services:
             - '@Akeneo\Pim\Enrichment\Category\API\Query\GetOwnedCategories'
             - '@Akeneo\Pim\Enrichment\Product\Domain\Query\GetNonViewableCategoryCodes'
             - '@Akeneo\Pim\Enrichment\Product\Domain\Query\GetCategoryCodes'
-            - '@Akeneo\Pim\Enrichment\Product\Domain\Query\GetProductUuids'
+            - '@Akeneo\Pim\Enrichment\Product\API\Query\GetProductUuids'
         tags:
             - { name: validator.constraint_validator }
