+++ src/Akeneo/Pim/Enrichment/Product/back/Test/Acceptance/Resources/config/services.yml
@@ -69,7 +69,7 @@ services:
     Akeneo\Pim\Enrichment\Product\Domain\Query\GetAttributeTypes:
         class: Akeneo\Pim\Enrichment\Product\Test\Acceptance\InMemory\InMemoryGetAttributeTypes
 
-    Akeneo\Pim\Enrichment\Product\Domain\Query\GetProductUuids:
+    Akeneo\Pim\Enrichment\Product\API\Query\GetProductUuids:
         class: Akeneo\Pim\Enrichment\Product\Test\Acceptance\InMemory\InMemoryGetProductUuids
         arguments:
             - '@pim_catalog.repository.product'
